requirements.txt
cognxcore==0.2.0
more_itertools==10.3.0
numba==0.59.1
numpy==2.0.0
PyAudio==0.2.14
python-dotenv==1.0.1
regex==2024.5.15
tiktoken==0.7.0
torch==2.3.1
torch_directml
tqdm==4.66.4
